+++ tencentcloud/clb/v20180317/models.go
@@ -5657,6 +5657,10 @@ type LoadBalancerTraffic struct {
 
 	// 最大出带宽，单位：Mbps
 	OutBandwidth *float64 `json:"OutBandwidth,omitempty" name:"OutBandwidth"`
+
+	// CLB域名
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Domain *string `json:"Domain,omitempty" name:"Domain"`
 }
 
 // Predefined struct for user