+++ tencentcloud/dcdb/v20180411/models.go
@@ -669,11 +669,7 @@ type CreateDCDBInstanceRequestParams struct {
 	// 虚拟私有网络子网 ID，VpcId不为空时必填
 	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
 
-	// 数据库引擎版本，当前可选：8.0.18，10.1.9，5.7.17。
-	// 8.0.18 - MySQL 8.0.18；
-	// 10.1.9 - Mariadb 10.1.9；
-	// 5.7.17 - Percona 5.7.17。
-	// 如果不填的话，默认为5.7.17，表示Percona 5.7.17。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 是否自动使用代金券进行支付，默认不使用。
@@ -747,11 +743,7 @@ type CreateDCDBInstanceRequest struct {
 	// 虚拟私有网络子网 ID，VpcId不为空时必填
 	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
 
-	// 数据库引擎版本，当前可选：8.0.18，10.1.9，5.7.17。
-	// 8.0.18 - MySQL 8.0.18；
-	// 10.1.9 - Mariadb 10.1.9；
-	// 5.7.17 - Percona 5.7.17。
-	// 如果不填的话，默认为5.7.17，表示Percona 5.7.17。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 是否自动使用代金券进行支付，默认不使用。
@@ -891,11 +883,7 @@ type CreateHourDCDBInstanceRequestParams struct {
 	//  查询实例规格获得。
 	ShardCpu *int64 `json:"ShardCpu,omitempty" name:"ShardCpu"`
 
-	// 数据库引擎版本，当前可选：10.0.10，10.1.9，5.7.17。
-	// 10.0.10 - Mariadb 10.0.10；
-	// 10.1.9 - Mariadb 10.1.9；
-	// 5.7.17 - Percona 5.7.17。
-	// 如果不填的话，默认为10.1.9，表示Mariadb 10.1.9。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 分片节点可用区分布，最多可填两个可用区。当分片规格为一主两从时，其中两个节点在第一个可用区。
@@ -966,11 +954,7 @@ type CreateHourDCDBInstanceRequest struct {
 	//  查询实例规格获得。
 	ShardCpu *int64 `json:"ShardCpu,omitempty" name:"ShardCpu"`
 
-	// 数据库引擎版本，当前可选：10.0.10，10.1.9，5.7.17。
-	// 10.0.10 - Mariadb 10.0.10；
-	// 10.1.9 - Mariadb 10.1.9；
-	// 5.7.17 - Percona 5.7.17。
-	// 如果不填的话，默认为10.1.9，表示Mariadb 10.1.9。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 分片节点可用区分布，最多可填两个可用区。当分片规格为一主两从时，其中两个节点在第一个可用区。