+++ tencentcloud/kms/v20190118/models.go
@@ -2403,6 +2403,10 @@ type GetServiceStatusResponseParams struct {
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	ExclusiveHSMEnabled *bool `json:"ExclusiveHSMEnabled,omitempty" name:"ExclusiveHSMEnabled"`
 
+	// KMS 订阅信息。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	SubscriptionInfo *string `json:"SubscriptionInfo,omitempty" name:"SubscriptionInfo"`
+
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
 }