+++ tencentcloud/mariadb/v20170312/models.go
@@ -647,7 +647,7 @@ type CreateDBInstanceRequestParams struct {
 	// 项目 ID，可以通过查看项目列表获取，不传则关联到默认项目
 	ProjectId *int64 `json:"ProjectId,omitempty" name:"ProjectId"`
 
-	// 数据库引擎版本，当前可选：8.0.18，10.1.9，5.7.17。如果不传的话，默认为 Percona 5.7.17。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 实例名称， 可以通过该字段自主的设置实例的名字
@@ -714,7 +714,7 @@ type CreateDBInstanceRequest struct {
 	// 项目 ID，可以通过查看项目列表获取，不传则关联到默认项目
 	ProjectId *int64 `json:"ProjectId,omitempty" name:"ProjectId"`
 
-	// 数据库引擎版本，当前可选：8.0.18，10.1.9，5.7.17。如果不传的话，默认为 Percona 5.7.17。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 实例名称， 可以通过该字段自主的设置实例的名字
@@ -1064,8 +1064,7 @@ type CreateHourDBInstanceRequestParams struct {
 	// 统一子网ID，VpcId有值时需填写
 	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
 
-	// 数据库引擎版本，当前可选：10.0.10，10.1.9，5.7.17。
-	// 如果不填的话，默认为10.1.9，表示Mariadb 10.1.9。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 自定义实例名称
@@ -1125,8 +1124,7 @@ type CreateHourDBInstanceRequest struct {
 	// 统一子网ID，VpcId有值时需填写
 	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
 
-	// 数据库引擎版本，当前可选：10.0.10，10.1.9，5.7.17。
-	// 如果不填的话，默认为10.1.9，表示Mariadb 10.1.9。
+	// 数据库引擎版本，当前可选：8.0，5.7，10.1，10.0。
 	DbVersionId *string `json:"DbVersionId,omitempty" name:"DbVersionId"`
 
 	// 自定义实例名称