+++ tencentcloud/trtc/v20190722/models.go
@@ -2444,6 +2444,14 @@ type McuLayoutVolume struct {
 
 	// SEI消息的payload_type，默认值100，取值范围100-254（244除外，244为我们内部自定义的时间戳SEI）
 	PayloadType *uint64 `json:"PayloadType,omitempty" name:"PayloadType"`
+
+	// SEI发送间隔，单位毫秒，默认值为1000。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Interval *uint64 `json:"Interval,omitempty" name:"Interval"`
+
+	// 取值范围[0,1]，填1：发送关键帧时会确保带SEI；填0：发送关键帧时不确保带SEI。默认值为0。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FollowIdr *uint64 `json:"FollowIdr,omitempty" name:"FollowIdr"`
 }
 
 type McuPassThrough struct {
@@ -2455,6 +2463,14 @@ type McuPassThrough struct {
 
 	// PayloadType为5，PayloadUuid必须填写。PayloadType不是5时，不需要填写，填写会被后台忽略。该值必须是32长度的十六进制。
 	PayloadUuid *string `json:"PayloadUuid,omitempty" name:"PayloadUuid"`
+
+	// SEI发送间隔，单位毫秒，默认值为1000。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Interval *uint64 `json:"Interval,omitempty" name:"Interval"`
+
+	// 取值范围[0,1]，填1：发送关键帧时会确保带SEI；填0：发送关键帧时不确保带SEI。默认值为0。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FollowIdr *uint64 `json:"FollowIdr,omitempty" name:"FollowIdr"`
 }
 
 type McuPublishCdnParam struct {